// FE/src/app/models/product.ts
export interface Product {
  productID: number;
  productName: string;
  productBrand: string;
  productScreen: string;
  operatingSystem: string;
  chip: string;
  RAM: string;
  battery: number;
  frontCamera: string;
  backCamera: string;
  productImage: string;
  quantityInStock: number;
  quantityBought: number;
  buyPrice: number;
}